fw/v0/usb_dfu: Only reboot on USB_RESET if we did something
From the DFU document it would look like we need to always reboot
on usb reset. But the fact is some host will issue resets ...

Also when passing the device to a VM it will disconnect/reconnect and
reset and it's annoying if the device goes back to runtime mode in
that case.

We always support detach command to force to go back to user mode.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>